+++ TerrariumGardenTech.Repositories/Entity/Category.cs
@@ -7,7 +7,7 @@ namespace TerrariumGardenTech.Repositories.Entity;
 
 public partial class Category
 {
-    public long CategoryId { get; set; }
+    public int CategoryId { get; set; }
 
     public string CategoryName { get; set; }
